requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-multipart
google-generativeai
python-dotenv